+++ pkg/epp/framework/plugins/scheduling/scorer/sessionaffinity/session_affinity.go
@@ -2,27 +2,19 @@ package sessionaffinity
 
 import (
 	"context"
-	"encoding/base64"
 	"encoding/json"
 	"fmt"
-	"strings"
 
-	"sigs.k8s.io/controller-runtime/pkg/log"
-
-	logutil "github.com/llm-d/llm-d-router/pkg/common/observability/logging"
 	"github.com/llm-d/llm-d-router/pkg/epp/framework/interface/datalayer"
 	"github.com/llm-d/llm-d-router/pkg/epp/framework/interface/plugin"
 	"github.com/llm-d/llm-d-router/pkg/epp/framework/interface/requestcontrol"
 	"github.com/llm-d/llm-d-router/pkg/epp/framework/interface/scheduling"
+	sessiontoken "github.com/llm-d/llm-d-router/pkg/epp/framework/plugins/scheduling/sessionaffinity"
 )
 
 const (
 	// SessionAffinityType is the type of the SessionAffinity scorer.
 	SessionAffinityType = "session-affinity-scorer"
-
-	// DefaultSessionTokenHeader is the default request/response header carrying
-	// the session token. It is shared by the session-affinity scorer and filter.
-	DefaultSessionTokenHeader = "x-session-token"
 )
 
 // parameters configures the SessionAffinity scorer.
@@ -30,55 +22,6 @@ type parameters struct {
 	// HeaderName overrides the default x-session-token header used to read and
 	// write the session token. When empty the default is used.
 	HeaderName string `json:"headerName"`
-}
-
-// NormalizeHeader lowercases and trims the configured session header name,
-// falling back to DefaultSessionTokenHeader when empty. It is shared by the
-// session-affinity scorer and filter.
-func NormalizeHeader(name string) string {
-	header := strings.ToLower(strings.TrimSpace(name))
-	if header == "" {
-		return DefaultSessionTokenHeader
-	}
-	return header
-}
-
-// DecodePodName decodes a base64-encoded session token into a pod
-// NamespacedName string. It returns "" when the token is empty or cannot be
-// decoded. It is shared by the session-affinity scorer and filter.
-func DecodePodName(ctx context.Context, token string) string {
-	if token == "" {
-		return ""
-	}
-	decoded, err := base64.StdEncoding.DecodeString(token)
-	if err != nil {
-		log.FromContext(ctx).Error(err, "Error decoding session header")
-		return ""
-	}
-	return string(decoded)
-}
-
-// WriteSessionResponseHeader encodes targetPod into sessionHeader on the
-// response sent to the client. It is shared by the session-affinity scorer and
-// filter; pluginType labels the originating plugin in logs.
-// TODO: this should be using a cookie and ensure not overriding any other
-// cookie values if present.
-// Tracked in https://github.com/llm-d/llm-d-router/issues/28
-func WriteSessionResponseHeader(ctx context.Context, pluginType, sessionHeader string, response *requestcontrol.Response, targetPod *datalayer.EndpointMetadata) {
-	if response == nil || targetPod == nil {
-		reqID := "undefined"
-		if response != nil {
-			reqID = response.RequestID
-		}
-		log.FromContext(ctx).V(logutil.DEBUG).Info("Session affinity - skip response header because response or targetPod is nil", "plugin", pluginType, "req id", reqID)
-		return
-	}
-
-	if response.Headers == nil { // TODO should always be populated?
-		response.Headers = make(map[string]string)
-	}
-
-	response.Headers[sessionHeader] = base64.StdEncoding.EncodeToString([]byte(targetPod.NamespacedName.String()))
 }
 
 // compile-time type assertion
@@ -102,7 +45,7 @@ func Factory(name string, rawParameters *json.Decoder, _ plugin.Handle) (plugin.
 func NewSessionAffinity(sessionHeader string) *SessionAffinity {
 	return &SessionAffinity{
 		typedName:     plugin.TypedName{Type: SessionAffinityType},
-		sessionHeader: NormalizeHeader(sessionHeader),
+		sessionHeader: sessiontoken.NormalizeHeader(sessionHeader),
 	}
 }
 
@@ -135,7 +78,7 @@ func (s *SessionAffinity) Category() scheduling.ScorerCategory {
 // Score assign a high score to the pod used in previous requests and zero to others
 func (s *SessionAffinity) Score(ctx context.Context, request *scheduling.InferenceRequest, endpoints []scheduling.Endpoint) map[scheduling.Endpoint]float64 {
 	scoredEndpoints := make(map[scheduling.Endpoint]float64)
-	podName := DecodePodName(ctx, request.Headers[s.sessionHeader])
+	podName := sessiontoken.DecodePodName(ctx, request.Headers[s.sessionHeader])
 
 	for _, endpoint := range endpoints {
 		scoredEndpoints[endpoint] = 0.0 // initial value
@@ -149,5 +92,5 @@ func (s *SessionAffinity) Score(ctx context.Context, request *scheduling.Inferen
 
 // ResponseHeader sets the session header on the response sent to the client.
 func (s *SessionAffinity) ResponseHeader(ctx context.Context, _ *scheduling.InferenceRequest, response *requestcontrol.Response, targetPod *datalayer.EndpointMetadata) {
-	WriteSessionResponseHeader(ctx, SessionAffinityType, s.sessionHeader, response, targetPod)
+	sessiontoken.WriteResponseHeader(ctx, SessionAffinityType, s.sessionHeader, response, targetPod)
 }
